timer in a separate file, skip bad files when training tokenizer
